+++ tests/Phake/Matchers/MethodMatcherTest.php
@@ -91,7 +91,7 @@ class MethodMatcherTest extends TestCase
      */
     public function testMatchesSuccessfullyMatches()
     {
-        Phake::when($this->rootArgumentMatcher)->doArgumentsMatch->thenReturn(true);
+        Phake::when($this->rootArgumentMatcher)->doArgumentsMatch(Phake::anyParameters())->thenReturn(true);
 
         $arguments = ['foo', 'bar'];
         $this->assertTrue($this->matcher->matches('foo', $arguments));
@@ -102,7 +102,7 @@ class MethodMatcherTest extends TestCase
      */
     public function testNoMatcherOnBadMethod()
     {
-        Phake::when($this->rootArgumentMatcher)->doArgumentsMatch->thenReturn(true);
+        Phake::when($this->rootArgumentMatcher)->doArgumentsMatch(Phake::anyParameters())->thenReturn(true);
 
         $arguments = ['foo', 'bar'];
         $this->assertFalse($this->matcher->matches('test', $arguments));
@@ -113,7 +113,7 @@ class MethodMatcherTest extends TestCase
      */
     public function testNoMatcherOnBadArg1()
     {
-        Phake::when($this->rootArgumentMatcher)->doArgumentsMatch->thenThrow(new Phake\Exception\MethodMatcherException());
+        Phake::when($this->rootArgumentMatcher)->doArgumentsMatch(Phake::anyParameters())->thenThrow(new Phake\Exception\MethodMatcherException());
 
         $arguments = ['foo', 'bar'];
         $this->assertFalse($this->matcher->matches('foo', $arguments));