+++ tests/Phake/Stubber/Answers/SmartDefaultAnswerTest.php
@@ -93,8 +93,7 @@ class SmartDefaultAnswerTest extends TestCase
         $context = new \PhakeTest_ScalarTypes();
         $cb = $this->answer->getAnswerCallback($context, 'callableReturn');
 
-        $this->assertEquals(function () {
-        }, $cb());
+        $this->assertTrue(is_callable($cb()));
     }
 
     public function testObjectReturn()