+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## Phake 4.7.0
+
+### Deprecation
+**Shorthand Stub and Verify notation are deprecated**
+Replace `Phake::when($mock)->foo by Phake::when($mock)->foo(Phake::anyParameters())` and `Phake::verify($mock)->foo` by `Phake::verify($mock)->foo(Phake::anyParameters())`.
+
+
 ## Phake 4.6.0
 
 ### New Feature